fix: styles

+++ backend/src/main/java/ca/bc/gov/backendstartapi/provider/OracleApiProvider.java
@@ -194,7 +194,7 @@ public class OracleApiProvider implements Provider {
   }
 
   @Override
-  public List<GeospatialOracleResDto> getPtGeospatialDataByIdList(List<Integer> ptIds) {
+  public List<GeospatialOracleResDto> getPtGeospatialDataByIdList(List<Long> ptIds) {
     String oracleApiUrl = String.format("%s/api/parent-trees/geospatial-data", rootUri);
 
     SparLog.info(
