+++ oracle-api/src/main/java/ca/bc/gov/backendstartapi/dto/GeospatialRespondDto.java
@@ -8,7 +8,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
         "This class represents a json response object when fetching geospatial data for a Parent"
             + " Tree.")
 public record GeospatialRespondDto(
-    @Schema(description = "The id of a parent tree", example = "8512") Integer parentTreeId,
+    @Schema(description = "The id of a parent tree", example = "8512") Long parentTreeId,
     @Schema(description = "Latitude degree", example = "49") Integer latitudeDegree,
     @Schema(description = "Latitude minute", example = "16") Integer latitudeMinute,
     @Schema(description = "Latitude second", example = "57") Integer latitudeSecond,
